backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
python-dotenv
alembic
psycopg[binary]
supabase
python-dateutil
rapidfuzz
unidecode
